requirements.txt
asgiref==3.4.1
codegen==1.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-extended-choices==1.3.3
django-filter==2.4.0
django-heroku==0.3.1
django-services==0.0.8
django-sms==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
environs==9.3.3
gunicorn==20.1.0
marshmallow==3.13.0
mysqlclient==2.0.3
npm==0.1.1
odict==1.7.0
optional-django==0.1.0
patterns==0.3
Pillow==8.3.2
plumber==1.6
psycopg2==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
services==0.1.1
six==1.16.0
sqlparse==0.4.1
sudo==1.0.0
transliterate==1.10.2
whitenoise==5.3.0
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
